requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-ninja==1.2.2
mysql-connector-python==9.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
